requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.6.1
async-timeout==4.0.2
attrs==22.1.0
charset-normalizer==2.1.1
click==8.1.3
defusedxml==0.7.1
Deprecated==1.2.13
frozenlist==1.3.1
h11==0.13.0
idna==3.3
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
multidict==6.0.2
odfpy==1.4.1
packaging==21.3
pydantic==1.10.2
pyparsing==3.0.9
python-ulid==1.1.0
pytz==2022.2.1
redis==4.3.4
sniffio==1.3.0
starlette==0.20.4
typing_extensions==4.3.0
ujson==5.4.0
uvicorn==0.18.3
uvloop==0.16.0
websockets==10.3
wrapt==1.14.1
yarl==1.8.1
